app/requirements-api.txt
flask
torch
torchvision
pillow
requests

